01/scripts/requirements.txt
numpy
pandas
scipy
names
python-codicefiscale
